requirements.txt
ekaPyCRUDApp==0.0.1
